+++ config/base/webpack.base.js
@@ -6,7 +6,7 @@ const webpack = require('webpack')
 const querystring = require('querystring')
 const extFs = require('yyl-fs')
 const es3ifyWebpackPlugin = require('es3ify-webpack-plugin')
-const UglifyjsWebpackPlugin = require('uglifyjs-webpack-plugin')
+const TerserWebpackPlugin = require('terser-webpack-plugin')
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
 
 const util = require('yyl-util')
@@ -197,8 +197,8 @@ const init = (config, iEnv) => {
     ],
     optimization: {
       minimizer: [
-        new UglifyjsWebpackPlugin({
-          uglifyOptions: {
+        new TerserWebpackPlugin({
+          terserOptions: {
             ie8: config.ie8 ? true : false,
             keep_fnames: config.keep_fnames ? true : false
           }